Revert "reworked all permission adding commands to use long instead of int to provide compatibility for icon ids"
This reverts commit ce8b50d6ad51f1c7baca0b319e11ee1ba1ed93a9.

+++ TS3-JAPI/src/main/java/net/sehales/ts3_japi/command/CmdChannelGroupAddPerm.java
@@ -12,17 +12,17 @@ public class CmdChannelGroupAddPerm extends Command {
         add(new KeyValueParameter("cgid", channelGroupId));
     }
 
-    public CmdChannelGroupAddPerm(int channelGroupId, int permId, long permValue) {
+    public CmdChannelGroupAddPerm(int channelGroupId, int permId, int permValue) {
         this(channelGroupId);
         addPerm(permId, permValue);
     }
 
-    public CmdChannelGroupAddPerm(int channelGroupId, String permsId, long permValue) {
+    public CmdChannelGroupAddPerm(int channelGroupId, String permsId, int permValue) {
         this(channelGroupId);
         addPerm(permsId, permValue);
     }
 
-    public CmdChannelGroupAddPerm addPerm(int permId, long permValue) {
+    public CmdChannelGroupAddPerm addPerm(int permId, int permValue) {
         array.add(new ListParameter(
                         new KeyValueParameter("permid", permId),
                         new KeyValueParameter("permvalue", permValue)
@@ -30,7 +30,7 @@ public class CmdChannelGroupAddPerm extends Command {
         return this;
     }
 
-    public CmdChannelGroupAddPerm addPerm(String permsId, long permValue) {
+    public CmdChannelGroupAddPerm addPerm(String permsId, int permValue) {
         array.add(new ListParameter(
                         new KeyValueParameter("permsid", permsId),
                         new KeyValueParameter("permvalue", permValue)
